Remove common files
